+++ examples/adfox.php
@@ -6,11 +6,14 @@ require_once dirname(__DIR__) . '/vendor/autoload.php';
 
 $feed = new Feed();
 
-// AdFox
-// Just use adNetwork method with empty ID attribute
-// You can use third optional attribute of adNetwork method for selecting id of turboContent element
-// in which Ad should placed
-// And don't forget to insert the AdFox banner code in the fourth attribute
+/*
+ AdFox
+ Just use adNetwork method with empty ID attribute.
+ You can use third optional attribute of adNetwork method for selecting id of turboContent element
+ in which Ad should placed (you also can use Content::adBlockPosition('') method for generate ad block position element
+ in any place of your content).
+ And don't forget to insert the AdFox banner code in the fourth attribute.
+ */
 
 $adFoxCode = '<div id="идентификатор контейнера"></div>
         <script>